Reorganize, synx and remove headers.
