--- SearchGitHubAPI/GitHub/GitHubAPIConditions.swift
@@ -1,20 +1,0 @@
-//
-//  GitHubAPIConditions.swift
-//  SearchGitHubAPI
-//
-//  Created by Michał Czerniakowski on 07.11.2016.
-//  Copyright © 2016 Michał Czerniakowski. All rights reserved.
-//
-
-import Foundation
-
-class GitHubAPIConditions
-
-{
-    var searchString: String?
-    var minStars = 0
-    
-    init() {
-        
-    }
-}
