--- SearchGitHubAPI/GitHub/GitHubAPIUsers.swift
@@ -1,14 +1,0 @@
-//
-//  GitHubAPIUsers.swift
-//  SearchGitHubAPI
-//
-//  Created by Michał Czerniakowski on 09.11.16.
-//  Copyright © 2016 Michał Czerniakowski. All rights reserved.
-//
-
-import Foundation
-
-class GitHubAPIUsers {
-    
-    var searchString: String?
-}
