--- SearchGitHubAPI/GitHub/MockDataProviderProtocol.swift
@@ -1,14 +1,0 @@
-//
-//  MockDataProviderProtocol.swift
-//  SearchGitHubAPI
-//
-//  Created by Michał Czerniakowski on 09.11.16.
-//  Copyright © 2016 Michał Czerniakowski. All rights reserved.
-//
-
-import Foundation
-
-@objc protocol MockDataProviderProtocol {
-    
-    func startSearch()
-}
