--- SearchGitHubAPI/Models/User.swift
@@ -1,17 +1,0 @@
-//
-//  User.swift
-//  SearchGitHubAPI
-//
-//  Created by Michał Czerniakowski on 09.11.16.
-//  Copyright © 2016 Michał Czerniakowski. All rights reserved.
-//
-
-import Foundation
-
-class User {
-    
-    var login: String?
-    var repoURL: String?
-    var avatarURL: String?
-    var score: Double = 0
-}
