+++ Sources/GitHub/GitHubAPIConditions.swift
@@ -1,0 +1,12 @@
+import Foundation
+
+class GitHubAPIConditions
+
+{
+    var searchString: String?
+    var minStars = 0
+    
+    init() {
+        
+    }
+}
