+++ Sources/GitHub/GitHubAPIUsers.swift
@@ -1,0 +1,6 @@
+import Foundation
+
+class GitHubAPIUsers {
+    
+    var searchString: String?
+}
