+++ Sources/GitHub/MockDataProviderProtocol.swift
@@ -1,0 +1,6 @@
+import Foundation
+
+@objc protocol MockDataProviderProtocol {
+    
+    func startSearch()
+}
