+++ Sources/Models/User.swift
@@ -1,0 +1,9 @@
+import Foundation
+
+class User {
+    
+    var login: String?
+    var repoURL: String?
+    var avatarURL: String?
+    var score: Double = 0
+}
