+++ Sources/Supporting Files/Assets.xcassets/Contents.json
