+++ Sources/Supporting Files/Assets.xcassets/star.imageset/Contents.json
