+++ Sources/Supporting Files/Assets.xcassets/star.imageset/star.png
