+++ Sources/Supporting Files/Info.plist
