+++ Sources/Views/CustomGITCell.swift
@@ -1,11 +1,3 @@
-//
-//  CustomGITCell.swift
-//  SearchGitHubAPI
-//
-//  Created by Michał Czerniakowski on 07.11.2016.
-//  Copyright © 2016 Michał Czerniakowski. All rights reserved.
-//
-
 import Foundation
 
 class CustomGITCell: UITableViewCell{
