+++ Sources/Views/DetailViewController.swift
@@ -1,11 +1,3 @@
-//
-//  DetailViewController.swift
-//  SearchGitHubAPI
-//
-//  Created by Michał Czerniakowski on 07.11.2016.
-//  Copyright © 2016 Michał Czerniakowski. All rights reserved.
-//
-
 import UIKit
 
 class DetailViewController: UIViewController {
