+++ Sources/Views/ViewController.swift
@@ -1,11 +1,3 @@
-//
-//  ViewController.swift
-//  SearchGitHubAPI
-//
-//  Created by Michał Czerniakowski on 04.11.2016.
-//  Copyright © 2016 Michał Czerniakowski. All rights reserved.
-//
-
 import UIKit
 
 class ViewController: UIViewController, UITableViewDataSource, UITableViewDelegate {
